+++ core/pkg/monitor/gpu_nvidia.go
@@ -1,4 +1,4 @@
-//go:build linux
+//go:build linux || windows
 
 package monitor
